requirements.txt
Django
djangorestframework
pytz
drf-yasg
pytest-django
pytest